Build Studio GUI on Windows

+++ libfive/include/libfive/tree/tree.hpp
@@ -112,7 +112,7 @@ public:
     static Tree nullary(Opcode::Opcode op);
 
     // Constructs a constant Tree with a floating-point value
-    Tree(float v);
+    Tree(double v);
 
     // Constructs a Tree from an OracleClause
     explicit Tree(std::unique_ptr<const OracleClause>&& oracle);
